requirements.txt
httpx>=0.27
python-telegram-bot>=21.0
anthropic>=0.25
supabase>=2.4
scikit-learn>=1.4
python-dotenv>=1.0
feedparser>=6.0
joblib>=1.3
